+++ src/game/main/GamePanel.java
@@ -5,9 +5,9 @@ import java.awt.*;
 import game.UI.GameOverScreen;
 import game.UI.Heart;
 import game.entities.Background;
-import game.entities.Dinosaur;
 import game.entities.Floor;
-import game.entities.Obstacle;
+import game.entities.character.Dinosaur;
+import game.entities.obstacles.Obstacle;
 import game.handlers.KeyHandler;
 import game.manager.CollisionManager;
 import game.manager.LivesManager;
